Refactor client_app.py into new directory structure
