+++ examples/federated-kaplan-meier-fitter/examplefkm/client_app.py
@@ -4,7 +4,8 @@ from typing import Dict, List, Tuple
 
 import flwr as fl
 import numpy as np
-from flwr.client import Client, ClientApp
+from flwr.client import Client
+from flwr.clientapp import ClientApp
 from flwr.common import Context, NDArray, NDArrays
 
 from examplefkm.task import load_partition
