+++ examples/quickstart-pandas/pandas_example/client_app.py
@@ -3,7 +3,7 @@
 import warnings
 
 import numpy as np
-from flwr.client import ClientApp
+from flwr.clientapp import ClientApp
 from flwr.common import Context, Message, MetricRecord, RecordDict
 from flwr_datasets import FederatedDataset
 from flwr_datasets.partitioner import IidPartitioner
