+++ framework/py/flwr/server/superlink/fleet/vce/vce_api.py
@@ -27,9 +27,9 @@ from typing import Callable, Optional
 from uuid import uuid4
 
 from flwr.app.error import Error
-from flwr.client.client_app import ClientApp, ClientAppException, LoadClientAppError
-from flwr.client.clientapp.utils import get_load_client_app_fn
+from flwr.clientapp.clientapp.utils import get_load_client_app_fn
 from flwr.client.run_info_store import DeprecatedRunInfoStore
+from flwr.clientapp.client_app import ClientApp, ClientAppException, LoadClientAppError
 from flwr.common import Message
 from flwr.common.constant import (
     HEARTBEAT_MAX_INTERVAL,
