+++ framework/py/flwr/simulation/ray_transport/ray_actor.py
@@ -24,7 +24,7 @@ import ray
 from ray import ObjectRef
 from ray.util.actor_pool import ActorPool
 
-from flwr.client.client_app import ClientApp, ClientAppException, LoadClientAppError
+from flwr.clientapp.client_app import ClientApp, ClientAppException, LoadClientAppError
 from flwr.common import Context, Message
 from flwr.common.logger import log
